requirements.txt
fastapi[all]
uvicorn[standard]
cx_Oracle
sqlalchemy
pandas
openpyxl
aioftp
pyzmq
loguru
chardet
gunicorn
